chat-ui/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
databricks-sdk>=0.20.0
